+++ tests/DeprecationAwareMiddlewareTest.php
@@ -7,7 +7,10 @@ namespace Wizofgoz\DeprecationLaravel\Tests;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Event;
 use Wizofgoz\DeprecationLaravel\Deprecated;
+use Wizofgoz\DeprecationLaravel\Events\DeprecatedResourceCalled;
+use Wizofgoz\DeprecationLaravel\Events\SunsettedResourceCalled;
 use Wizofgoz\DeprecationLaravel\Http\Middleware\DeprecationAwareMiddleware;
 use Wizofgoz\DeprecationLaravel\Links\AlternateLink;
 use Wizofgoz\DeprecationLaravel\Links\Link;
@@ -90,5 +93,41 @@ class DeprecationAwareMiddlewareTest extends TestCase
         $this->assertFalse($response->headers->has(Deprecated::HEADER));
         $this->assertFalse($response->headers->has(Link::HEADER));
         $this->assertFalse($response->headers->has(Sunsetted::HEADER));
+    }
+
+    /** @test */
+    public function it_emits_an_event_when_attaching_deprecated_header()
+    {
+        Event::fake();
+
+        $middleware = $this->createMiddleware();
+        $deprecated = Deprecated::new();
+        $this->getService()->deprecate($deprecated);
+
+        $middleware->handle(new Request(), fn () => new Response());
+
+        Event::assertDispatched(
+            DeprecatedResourceCalled::class,
+            fn (DeprecatedResourceCalled $event) => $event->deprecation === $deprecated
+        );
+        Event::assertDispatchedTimes(DeprecatedResourceCalled::class, 1);
+    }
+
+    /** @test */
+    public function it_emits_an_event_when_attaching_sunsetted_header()
+    {
+        Event::fake();
+
+        $middleware = $this->createMiddleware();
+        $sunsetted = Sunsetted::new(Carbon::now());
+        $this->getService()->sunset($sunsetted);
+
+        $middleware->handle(new Request(), fn () => new Response());
+
+        Event::assertDispatched(
+            SunsettedResourceCalled::class,
+            fn (SunsettedResourceCalled $event) => $event->sunset === $sunsetted
+        );
+        Event::assertDispatchedTimes(SunsettedResourceCalled::class, 1);
     }
 }