+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to `deprecation-laravel` will be documented in this file.
 
+## 1.1.0 - 2021-04-14
+ - Fixed typo in header name
+ - Added Laravel events when headers are added to the response
+ - Added code coverage stats
+
 ## 1.0.0 - 2021-04-13
 
 - initial release